requirements.txt
pyOpenSSL >= 16.0.0
